requirements.txt
# please refer to the Dockerfile if you are familiar with building environment using Docker
# for data generation
albumentations
Cython
matplotlib==2.2.4
numpy==1.16.4
opencv-contrib-python
pandas
Pillow
PyYAML
scikit-image
scipy
sklearn

# mainly for DeepLab
jupyter
prettytable
tensorflow-gpu==1.12.0

# mainly for DeepCrack
dominate==2.3.1
torch==0.4.1
torchvision==0.2.1
visdom==0.1.7